+++ OplusUpdater/cmd/updater/main_test.go
@@ -53,6 +53,7 @@ func TestQueryArgsFromCommandAdvancedFlags(t *testing.T) {
 		"--guid", validGUID,
 		"--anti",
 		"--gray",
+		"--graynew",
 		"--pre",
 		"--language", "en-IN",
 		"--rom-version", "RMX3301_15.0.0.1410(EX01)",
@@ -78,6 +79,9 @@ func TestQueryArgsFromCommandAdvancedFlags(t *testing.T) {
 	}
 	if !args.Gray {
 		t.Fatal("expected gray flag to be true")
+	}
+	if !args.GrayNew {
+		t.Fatal("expected graynew flag to be true")
 	}
 	if !args.Pre {
 		t.Fatal("expected pre flag to be true")